Document

+++ vertx-pin/zero-ui/src/main/resources/plugin/sql/ui/UI_FORM.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS UI_FORM
     `KEY`        VARCHAR(36) COMMENT '「key」- 主键',
     `NAME`       VARCHAR(255) COMMENT '「name」- 表单名称',
     `CODE`       VARCHAR(255) COMMENT '「code」- 表单系统编码',
-    `WINDOW`     INTEGER COMMENT '「window」- window, Form对应的窗口配置',
+    `WINDOW`     DECIMAL(10, 2) COMMENT '「window」- window, Form对应的窗口配置',
     `COLUMNS`    INTEGER COMMENT '「columns」- columns, Form对应的配置',
     `HIDDEN`     TEXT COMMENT '「hidden」- 隐藏字段专用配置',
     `ROW`        TEXT COMMENT '「rowConfig/rowClass」- 行专用配置',
